[LT-71] Refactor data type from double to big decimal

+++ src/main/java/com/lankatrails/lankatrails_backend/dtos/request/PriceConfigDTO.java
@@ -1,7 +1,14 @@
 package com.lankatrails.lankatrails_backend.dtos.request;
 
+import java.math.BigDecimal;
+
 import com.lankatrails.lankatrails_backend.model.enums.PriceType;
-import lombok.*;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 @Getter
 @Setter
@@ -9,18 +16,18 @@ import lombok.*;
 @AllArgsConstructor
 @Builder
 public class PriceConfigDTO {
-    private Double fixedPrice;
-    private Double pricePerUnit;
-    private Double pricePerAdult;
-    private Double pricePerChild;
+    private BigDecimal fixedPrice;
+    private BigDecimal pricePerUnit;
+    private BigDecimal pricePerAdult;
+    private BigDecimal pricePerChild;
     private PriceType priceType; // Use String to represent enum
-    private Double extraChargePerUnit;
-    private Double extraPerAdult;
-    private Double extraPerChild;
+    private BigDecimal extraChargePerUnit;
+    private BigDecimal extraPerAdult;
+    private BigDecimal extraPerChild;
     private PriceType extraChargeType; // Use String to represent enum
     private Boolean allowAdvancePayment;
-    private Double advancePaymentPercentage;
-    private Double advancePaymentFixedAmount;
+    private BigDecimal advancePaymentPercentage;
+    private BigDecimal advancePaymentFixedAmount;
     private Boolean requiresDeposit;
-    private Double depositAmount;
+    private BigDecimal depositAmount;
 }
